sys/src/9/port: use uintptr for the kernel malloc family to match the userland libraries

+++ sys/src/9/port/lib.h
@@ -9,7 +9,7 @@
  * mem routines
  */
 extern	void*	memccpy(void*, void*, int, ulong);
-extern	void*	memset(void*, int, ulong);
+extern	void*	memset(void*, int, uintptr);
 extern	int	memcmp(void*, void*, ulong);
 extern	void*	memmove(void*, void*, ulong);
 extern	void*	memchr(void*, int, ulong);
